@@ -1,3 +1,5 @@
 module github.com/areYouLazy/libhosty
 
-go 1.16
+go 1.21
+
+require golang.org/x/exp v0.0.0-20240808152545-0cdaa3abc0fa
