@@ -1,19 +1,13 @@
-//Package libhosty is a pure golang library to manipulate the hosts file
+// Package libhosty is a pure golang library to manipulate the hosts file
 package libhosty
 
 import (
-	"io/ioutil"
 	"net"
-	"os"
 	"regexp"
-	"runtime"
 	"strings"
 	"sync"
-)
 
-const (
-	//Version exposes library version
-	Version = "2.0"
+	"golang.org/x/exp/slices"
 )
 
 const (
@@ -27,7 +21,7 @@ const (
 	hostsFileName = "hosts"
 )
 
-//LineType define a safe type for line type enumeration
+// LineType define a safe type for line type enumeration
 type LineType int
 
 const (
@@ -44,25 +38,26 @@ const (
 	LineTypeAddress LineType = 30
 )
 
-//HostsFileConfig defines parameters to find hosts file.
-// FilePath is the absolute path of the hosts file (filename included)
-type HostsFileConfig struct {
-	FilePath string
+func (lt LineType) String() string {
+	switch lt {
+	case LineTypeEmpty:
+		return "line-type-empty"
+	case LineTypeComment:
+		return "line-type-comment"
+	case LineTypeAddress:
+		return "line-type-address"
+	default:
+		return "line-type-unknown"
+	}
 }
 
-//HostsFileLine holds hosts file lines data
+// HostsFileLine holds hosts file lines data
 type HostsFileLine struct {
-	//Number is the original line number
-	Number int
-
 	//LineType defines the line type
 	Type LineType
 
 	//Address is a net.IP representation of the address
 	Address net.IP
-
-	//Parts is a slice of the line splitted by '#'
-	Parts []string
 
 	//Hostnames is a slice of hostnames for the relative IP
 	Hostnames []string
@@ -75,59 +70,26 @@ type HostsFileLine struct {
 
 	//IsCommented to know if the current ADDRESS line is commented out (starts with '#')
 	IsCommented bool
-
-	//trimed is a trimed version (no spaces before and after) of the line
-	trimed string
 }
 
-//HostsFile is a reference for the hosts file configuration and lines
+// HostsFile is a reference for the hosts file configuration and lines
 type HostsFile struct {
 	sync.Mutex
 
-	//Config reference to a HostsConfig object
-	Config *HostsFileConfig
+	//Path is the file path and name
+	Path string
 
 	//HostsFileLines slice of HostsFileLine objects
 	HostsFileLines []HostsFileLine
 }
 
-//InitWithConfig returns a new instance of a hostsfile.
-// InitWithConfig is meant to be used with a custom conf file
-// however InitWithConfig() will fallback to Init() if conf is nill
-// You should use Init() to load hosts file from default location
-func InitWithConfig(conf *HostsFileConfig) (*HostsFile, error) {
-	var config *HostsFileConfig
-	var err error
-
-	if conf != nil {
-		config = conf
-	} else {
-		return Init()
-	}
-
-	// allocate a new HostsFile object
-	hf := &HostsFile{
-		// use default configuration
-		Config: config,
-
-		// allocate a new slice of HostsFileLine objects
-		HostsFileLines: make([]HostsFileLine, 0),
-	}
-
-	// parse the hosts file and load file lines
-	hf.HostsFileLines, err = ParseHostsFile(hf.Config.FilePath)
-	if err != nil {
-		return nil, err
-	}
-
-	//return HostsFile
-	return hf, nil
-}
-
-//Init returns a new instance of a hostsfile.
+// Init returns a new instance of a hostsfile.
 func Init() (*HostsFile, error) {
-	// initialize hostsConfig
-	config, err := NewHostsFileConfig("")
+	// get hosts file default path
+	fpath := GetOSHostsFilePath()
+
+	// parse hosts file from default path
+	hostsFileLines, err := ParseHostsFile(fpath)
 	if err != nil {
 		return nil, err
 	}
@@ -135,56 +97,57 @@ func Init() (*HostsFile, error) {
 	// allocate a new HostsFile object
 	hf := &HostsFile{
 		// use default configuration
-		Config: config,
+		Path: fpath,
 
 		// allocate a new slice of HostsFileLine objects
-		HostsFileLines: make([]HostsFileLine, 0),
-	}
-
-	// parse the hosts file and load file lines
-	hf.HostsFileLines, err = ParseHostsFile(hf.Config.FilePath)
-	if err != nil {
-		return nil, err
+		HostsFileLines: hostsFileLines,
 	}
 
 	//return HostsFile
 	return hf, nil
 }
 
-//NewHostsFileConfig loads hosts file based on environment.
-// NewHostsFileConfig initialize the default file path based
-// on the OS or from a given location if a custom path is provided
-func NewHostsFileConfig(path string) (*HostsFileConfig, error) {
-	// allocate hostsConfig
-	var hc *HostsFileConfig
-
-	// ensure custom path exists
-	// https://stackoverflow.com/questions/12518876/how-to-check-if-a-file-exists-in-go
-	if fh, err := os.Stat(path); err == nil {
-		// eusure custom path points to a file (not a directory)
-		if !fh.IsDir() {
-			hc = &HostsFileConfig{
-				FilePath: path,
-			}
-		}
-	} else {
-		// check os to construct default path
-		switch runtime.GOOS {
-		case "windows":
-			hc = &HostsFileConfig{
-				FilePath: windowsFilePath + hostsFileName,
-			}
-		default:
-			hc = &HostsFileConfig{
-				FilePath: unixFilePath + hostsFileName,
-			}
-		}
+func InitFromCustomPath(path string) (*HostsFile, error) {
+	// parse hosts file from default path
+	hostsFileLines, err := ParseHostsFile(path)
+	if err != nil {
+		return nil, err
 	}
 
-	return hc, nil
+	// allocate a new HostsFile object
+	hf := &HostsFile{
+		// use default configuration
+		Path: path,
+
+		// allocate a new slice of HostsFileLine objects
+		HostsFileLines: hostsFileLines,
+	}
+
+	//return HostsFile
+	return hf, nil
 }
 
-//GetHostsFileLines returns every address row
+func InitFromString(lines string) (*HostsFile, error) {
+	// parse inline hosts file
+	hostsFileLines, err := ParseHostsFileFromString(lines)
+	if err != nil {
+		return nil, err
+	}
+
+	// allocate a new HostsFile object
+	hf := &HostsFile{
+		// use default configuration
+		Path: "",
+
+		// allocate a new slice of HostsFileLine objects
+		HostsFileLines: hostsFileLines,
+	}
+
+	//return HostsFile
+	return hf, nil
+}
+
+// GetHostsFileLines returns every address row
 func (h *HostsFile) GetHostsFileLines() []*HostsFileLine {
 	var hfl []*HostsFileLine
 
@@ -197,26 +160,12 @@ func (h *HostsFile) GetHostsFileLines() []*HostsFileLine {
 	return hfl
 }
 
-//GetHostsFileLineByRow returns a ponter to the given HostsFileLine row
+// GetHostsFileLineByRow returns a ponter to the given HostsFileLine row
 func (h *HostsFile) GetHostsFileLineByRow(row int) *HostsFileLine {
 	return &h.HostsFileLines[row]
 }
 
-//GetHostsFileLineByIP returns the index of the line and a ponter to the given HostsFileLine line
-func (h *HostsFile) GetHostsFileLineByIP(ip net.IP) (int, *HostsFileLine) {
-	if ip == nil {
-		return -1, nil
-	}
-
-	for idx := range h.HostsFileLines {
-		if net.IP.Equal(ip, h.HostsFileLines[idx].Address) {
-			return idx, &h.HostsFileLines[idx]
-		}
-	}
-
-	return -1, nil
-}
-
+// GetHostsFileLinesByIP returns every line that maches a given IP
 func (h *HostsFile) GetHostsFileLinesByIP(ip net.IP) []*HostsFileLine {
 	if ip == nil {
 		return nil
@@ -233,30 +182,13 @@ func (h *HostsFile) GetHostsFileLinesByIP(ip net.IP) []*HostsFileLine {
 	return hfl
 }
 
-//GetHostsFileLineByAddress returns the index of the line and a ponter to the given HostsFileLine line
-func (h *HostsFile) GetHostsFileLineByAddress(address string) (int, *HostsFileLine) {
-	ip := net.ParseIP(address)
-	return h.GetHostsFileLineByIP(ip)
-}
-
+// GetHostsFileLinesByAddress returns every line that maches a given IP as String
 func (h *HostsFile) GetHostsFileLinesByAddress(address string) []*HostsFileLine {
 	ip := net.ParseIP(address)
 	return h.GetHostsFileLinesByIP(ip)
 }
 
-//GetHostsFileLineByHostname returns the index of the line and a ponter to the given HostsFileLine line
-func (h *HostsFile) GetHostsFileLineByHostname(hostname string) (int, *HostsFileLine) {
-	for idx := range h.HostsFileLines {
-		for _, hn := range h.HostsFileLines[idx].Hostnames {
-			if hn == hostname {
-				return idx, &h.HostsFileLines[idx]
-			}
-		}
-	}
-
-	return -1, nil
-}
-
+// GetHostsFileLinesByHostname returns every line that maches a given Hostname
 func (h *HostsFile) GetHostsFileLinesByHostname(hostname string) []*HostsFileLine {
 	hfl := make([]*HostsFileLine, 0)
 
@@ -272,15 +204,16 @@ func (h *HostsFile) GetHostsFileLinesByHostname(hostname string) []*HostsFileLin
 	return hfl
 }
 
-func (h *HostsFile) GetHostsFileLinesByHostnameAsRegexp(hostname string) []*HostsFileLine {
-	hfl := make([]*HostsFileLine, 0)
+// GetHostsFileLinesByRegexp returns every line that maches a given regexp
+func (h *HostsFile) GetHostsFileLinesByRegexp(pattern string) []HostsFileLine {
+	hfl := make([]HostsFileLine, 0)
 
-	reg := regexp.MustCompile(hostname)
+	reg := regexp.MustCompile(pattern)
 
 	for idx := range h.HostsFileLines {
 		for _, hn := range h.HostsFileLines[idx].Hostnames {
 			if reg.MatchString(hn) {
-				hfl = append(hfl, &h.HostsFileLines[idx])
+				hfl = append(hfl, h.HostsFileLines[idx])
 				continue
 			}
 		}
@@ -289,53 +222,7 @@ func (h *HostsFile) GetHostsFileLinesByHostnameAsRegexp(hostname string) []*Host
 	return hfl
 }
 
-//RenderHostsFile render and returns the hosts file with the lineFormatter() routine
-func (h *HostsFile) RenderHostsFile() string {
-	// allocate a buffer for file lines
-	var sliceBuffer []string
-
-	// iterate HostsFileLines and popolate the buffer with formatted lines
-	for _, l := range h.HostsFileLines {
-		sliceBuffer = append(sliceBuffer, lineFormatter(l))
-	}
-
-	// strings.Join() prevent the last line from being a new blank line
-	// as opposite to a for loop with fmt.Printf(buffer + '\n')
-	return strings.Join(sliceBuffer, "\n")
-}
-
-//RenderHostsFileLine render and returns the given hosts line with the lineFormatter() routine
-func (h *HostsFile) RenderHostsFileLine(row int) string {
-	// iterate to find the row to render
-	if len(h.HostsFileLines) > row {
-		return lineFormatter(h.HostsFileLines[row])
-	}
-
-	return ""
-}
-
-//SaveHostsFile write hosts file to configured path.
-// error is not nil if something goes wrong
-func (h *HostsFile) SaveHostsFile() error {
-	return h.SaveHostsFileAs(h.Config.FilePath)
-}
-
-//SaveHostsFileAs write hosts file to the given path.
-// error is not nil if something goes wrong
-func (h *HostsFile) SaveHostsFileAs(path string) error {
-	// render the file as a byte slice
-	dataBytes := []byte(h.RenderHostsFile())
-
-	// write file to disk
-	err := ioutil.WriteFile(path, dataBytes, 0644)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-//RemoveHostsFileLineByRow remove row at given index from HostsFileLines
+// RemoveHostsFileLineByRow remove row at given index from HostsFileLines
 func (h *HostsFile) RemoveHostsFileLineByRow(row int) {
 	// prevent out-of-index
 	if row < len(h.HostsFileLines) {
@@ -345,15 +232,7 @@ func (h *HostsFile) RemoveHostsFileLineByRow(row int) {
 	}
 }
 
-func (h *HostsFile) RemoveHostsFileLineByIP(ip net.IP) {
-	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
-		if net.IP.Equal(ip, h.HostsFileLines[idx].Address) {
-			h.RemoveHostsFileLineByRow(idx)
-			return
-		}
-	}
-}
-
+// RemoveHostFileLinesByIP remove every line that matches a given IP
 func (h *HostsFile) RemoveHostsFileLinesByIP(ip net.IP) {
 	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
 		if net.IP.Equal(ip, h.HostsFileLines[idx].Address) {
@@ -362,44 +241,14 @@ func (h *HostsFile) RemoveHostsFileLinesByIP(ip net.IP) {
 	}
 }
 
-func (h *HostsFile) RemoveHostsFileLineByAddress(address string) {
-	ip := net.ParseIP(address)
-
-	h.RemoveHostsFileLineByIP(ip)
-}
-
+// RemoveHostFileLinesByAddress remove every line that matches a given IP as String
 func (h *HostsFile) RemoveHostsFileLinesByAddress(address string) {
 	ip := net.ParseIP(address)
 
 	h.RemoveHostsFileLinesByIP(ip)
 }
 
-func (h *HostsFile) RemoveHostsFileLineByHostname(hostname string) {
-	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
-		if h.HostsFileLines[idx].Type == LineTypeAddress {
-			for _, hn := range h.HostsFileLines[idx].Hostnames {
-				if hn == hostname {
-					h.RemoveHostsFileLineByRow(idx)
-					return
-				}
-			}
-		}
-	}
-}
-
-func (h *HostsFile) RemoveHostsFileLinesByHostnameAsRegexp(hostname string) {
-	reg := regexp.MustCompile(hostname)
-
-	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
-		for _, hn := range h.HostsFileLines[idx].Hostnames {
-			if reg.MatchString(hn) {
-				h.RemoveHostsFileLineByRow(idx)
-				continue
-			}
-		}
-	}
-}
-
+// RemoveHostFileLinesByHostname remove every line that matches a given Hostname
 func (h *HostsFile) RemoveHostsFileLinesByHostname(hostname string) {
 	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
 		if h.HostsFileLines[idx].Type == LineTypeAddress {
@@ -413,7 +262,21 @@ func (h *HostsFile) RemoveHostsFileLinesByHostname(hostname string) {
 	}
 }
 
-//LookupByHostname check if the given fqdn exists.
+// RemoveHostFileLinesByRegexp remove every line that matches a given regexp
+func (h *HostsFile) RemoveHostsFileLinesByRegexp(pattern string) {
+	reg := regexp.MustCompile(pattern)
+
+	for idx := len(h.HostsFileLines) - 1; idx >= 0; idx-- {
+		for _, hn := range h.HostsFileLines[idx].Hostnames {
+			if reg.MatchString(hn) {
+				h.RemoveHostsFileLineByRow(idx)
+				continue
+			}
+		}
+	}
+}
+
+// LookupByHostname check if the given fqdn exists.
 // if yes, it returns the index of the address and the associated address.
 // error is not nil if something goes wrong
 func (h *HostsFile) LookupByHostname(hostname string) (int, net.IP, error) {
@@ -428,7 +291,7 @@ func (h *HostsFile) LookupByHostname(hostname string) (int, net.IP, error) {
 	return -1, nil, ErrHostnameNotFound
 }
 
-//AddHostsFileLineRaw add the given ip/fqdn/comment pair
+// AddHostsFileLineRaw add the given ip/fqdn/comment pair
 // this is different from AddHostFileLine because it does not take care of duplicates
 // this just append the new entry to the hosts file
 func (h *HostsFile) AddHostsFileLineRaw(ipRaw, fqdnRaw, comment string) (int, *HostsFileLine, error) {
@@ -436,6 +299,9 @@ func (h *HostsFile) AddHostsFileLineRaw(ipRaw, fqdnRaw, comment string) (int, *H
 	hostname := strings.ToLower(fqdnRaw)
 	// parse ip to net.IP
 	ip := net.ParseIP(ipRaw)
+
+	// get index
+	idx := len(h.HostsFileLines)
 
 	// if we have a valid IP
 	if ip != nil {
@@ -451,18 +317,15 @@ func (h *HostsFile) AddHostsFileLineRaw(ipRaw, fqdnRaw, comment string) (int, *H
 		// append to hosts
 		h.HostsFileLines = append(h.HostsFileLines, hfl)
 
-		// get index
-		idx := len(h.HostsFileLines) - 1
-
 		// return created entry
-		return idx, &h.HostsFileLines[idx], nil
+		return idx, &hfl, nil
 	}
 
 	// return error
 	return -1, nil, ErrCannotParseIPAddress(ipRaw)
 }
 
-//AddHostsFileLine add the given ip/fqdn/comment pair, cleanup is done for previous entry.
+// AddHostsFileLine add the given ip/fqdn/comment pair, cleanup is done for previous entry.
 // it returns the index of the edited (created) line and a pointer to the hostsfileline object.
 // error is not nil if something goes wrong
 func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *HostsFileLine, error) {
@@ -473,9 +336,9 @@ func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *Host
 
 	// if we have a valid IP
 	if ip != nil {
-		//check if we alredy have the fqdn
+		// check if we alredy have the fqdn
 		if idx, addr, err := h.LookupByHostname(hostname); err == nil {
-			//if actual ip is the same as the given one, we are done
+			// if actual ip is the same as the given one, we are done
 			if net.IP.Equal(addr, ip) {
 				// handle comment
 				if comment != "" {
@@ -485,26 +348,42 @@ func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *Host
 				return idx, &h.HostsFileLines[idx], nil
 			}
 
-			//if address is different, we need to remove the hostname from the previous entry
+			// if address is different, we need to remove the hostname from the previous entry
 			for hostIdx, hn := range h.HostsFileLines[idx].Hostnames {
+				// if hostnames matches
 				if hn == hostname {
+					// remove hostname if there's at least one another hostname
 					if len(h.HostsFileLines[idx].Hostnames) > 1 {
 						h.Lock()
 						h.HostsFileLines[idx].Hostnames = append(h.HostsFileLines[idx].Hostnames[:hostIdx], h.HostsFileLines[idx].Hostnames[hostIdx+1:]...)
 						h.Unlock()
-					}
-
-					//remove the line if there are no more hostnames (other than the actual one)
-					if len(h.HostsFileLines[idx].Hostnames) <= 1 {
+					} else {
+						// remove the whole line
 						h.RemoveHostsFileLineByRow(idx)
 					}
 				}
 			}
 		}
 
-		//if we alredy have the address, just add the hostname to that line
+		// index saves the last matching index for the next for loop
+		// in this way, if we find a matching line (same IP)
+		// but needs to create a new line (hostname limit exceeded)
+		// it would be nice to place the new line next to the existing one
+		index := -1
+
+		// we don't have the fqdn
+		// if we already have the address, just add the hostname to that line
 		for idx, hfl := range h.HostsFileLines {
 			if net.IP.Equal(hfl.Address, ip) {
+				// if we already have 6 hostnames, just continue
+				// we'll either find another matching line
+				// or we'll end up creating a new line
+				if len(h.HostsFileLines[idx].Hostnames) >= 6 {
+					// save index
+					index = idx
+					continue
+				}
+
 				h.Lock()
 				h.HostsFileLines[idx].Hostnames = append(h.HostsFileLines[idx].Hostnames, hostname)
 				h.Unlock()
@@ -520,11 +399,21 @@ func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *Host
 			}
 		}
 
+		var idx int
+		// we found a matching line
+		if index != -1 {
+			// so set the index next to the matching line
+			idx = index + 1
+		} else {
+			idx = len(h.HostsFileLines)
+		}
+
 		// at this point we need to create new host line
 		hfl := HostsFileLine{
 			Type:        LineTypeAddress,
 			Address:     ip,
 			Hostnames:   []string{hostname},
+			Raw:         "",
 			Comment:     comment,
 			IsCommented: false,
 		}
@@ -532,11 +421,14 @@ func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *Host
 		// generate raw version of the line
 		hfl.Raw = lineFormatter(hfl)
 
-		// append to hosts
-		h.HostsFileLines = append(h.HostsFileLines, hfl)
-
-		// get index
-		idx := len(h.HostsFileLines) - 1
+		// if we found a matching line (index != -1)
+		// append the new line after the matched one
+		if index != -1 {
+			h.HostsFileLines = slices.Insert(h.HostsFileLines, idx, hfl)
+		} else {
+			// else append to hosts
+			h.HostsFileLines = append(h.HostsFileLines, hfl)
+		}
 
 		// return created entry
 		return idx, &h.HostsFileLines[idx], nil
@@ -546,44 +438,53 @@ func (h *HostsFile) AddHostsFileLine(ipRaw, fqdnRaw, comment string) (int, *Host
 	return -1, nil, ErrCannotParseIPAddress(ipRaw)
 }
 
-//AddCommentFileLine adds a new line of type comment with the given comment.
+// AddCommentFileLine adds a new line of type comment with the given comment.
 // it returns the index of the edited (created) line and a pointer to the hostsfileline object.
 // error is not nil if something goes wrong
 func (h *HostsFile) AddCommentFileLine(comment string) (int, *HostsFileLine, error) {
 	h.Lock()
 	defer h.Unlock()
 
+	idx := len(h.HostsFileLines)
+
 	hfl := HostsFileLine{
-		Type:    LineTypeComment,
-		Raw:     "# " + comment,
-		Comment: comment,
+		Type:        LineTypeComment,
+		Address:     []byte{},
+		Hostnames:   []string{},
+		Raw:         "# " + comment,
+		Comment:     comment,
+		IsCommented: false,
 	}
 
 	hfl.Raw = lineFormatter(hfl)
 
 	h.HostsFileLines = append(h.HostsFileLines, hfl)
-	idx := len(h.HostsFileLines) - 1
 	return idx, &h.HostsFileLines[idx], nil
 }
 
-//AddEmptyFileLine adds a new line of type empty.
+// AddEmptyFileLine adds a new line of type empty.
 // it returns the index of the edited (created) line and a pointer to the hostsfileline object.
 // error is not nil if something goes wrong
 func (h *HostsFile) AddEmptyFileLine() (int, *HostsFileLine, error) {
 	h.Lock()
 	defer h.Unlock()
 
+	idx := len(h.HostsFileLines)
+
 	hfl := HostsFileLine{
-		Type: LineTypeEmpty,
-		Raw:  "",
+		Type:        LineTypeEmpty,
+		Address:     []byte{},
+		Hostnames:   []string{},
+		Raw:         "",
+		Comment:     "",
+		IsCommented: false,
 	}
 
 	h.HostsFileLines = append(h.HostsFileLines, hfl)
-	idx := len(h.HostsFileLines) - 1
 	return idx, &h.HostsFileLines[idx], nil
 }
 
-//CommentHostsFileLineByRow set the IsCommented bit for the given row to true
+// CommentHostsFileLineByRow set the IsCommented bit for the given row to true
 func (h *HostsFile) CommentHostsFileLineByRow(row int) error {
 	h.Lock()
 	defer h.Unlock()
@@ -606,27 +507,7 @@ func (h *HostsFile) CommentHostsFileLineByRow(row int) error {
 	return ErrUnknown
 }
 
-//CommentHostsFileLineByIP set the IsCommented bit for the given address to true
-func (h *HostsFile) CommentHostsFileLineByIP(ip net.IP) error {
-	h.Lock()
-	defer h.Unlock()
-
-	for idx := range h.HostsFileLines {
-		if net.IP.Equal(ip, h.HostsFileLines[idx].Address) {
-			if !h.HostsFileLines[idx].IsCommented {
-				h.HostsFileLines[idx].IsCommented = true
-
-				h.HostsFileLines[idx].Raw = h.RenderHostsFileLine(idx)
-				return nil
-			}
-
-			return ErrAlredyCommentedLine
-		}
-	}
-
-	return ErrAddressNotFound
-}
-
+// CommentHostsFileLinesByIP set IsCommented to true on every line that matches a given net.IP
 func (h *HostsFile) CommentHostsFileLinesByIP(ip net.IP) {
 	h.Lock()
 	defer h.Unlock()
@@ -642,41 +523,14 @@ func (h *HostsFile) CommentHostsFileLinesByIP(ip net.IP) {
 	}
 }
 
-//CommentHostsFileLineByAddress set the IsCommented bit for the given address as string to true
-func (h *HostsFile) CommentHostsFileLineByAddress(address string) error {
-	ip := net.ParseIP(address)
-
-	return h.CommentHostsFileLineByIP(ip)
-}
-
+// CommentHostsFileLinesByAddress set IsCommented to true on every line that matches a given net.IP
 func (h *HostsFile) CommentHostsFileLinesByAddress(address string) {
 	ip := net.ParseIP(address)
+
 	h.CommentHostsFileLinesByIP(ip)
 }
 
-//CommentHostsFileLineByHostname set the IsCommented bit for the given hostname to true
-func (h *HostsFile) CommentHostsFileLineByHostname(hostname string) error {
-	h.Lock()
-	defer h.Unlock()
-
-	for idx := range h.HostsFileLines {
-		for _, hn := range h.HostsFileLines[idx].Hostnames {
-			if hn == hostname {
-				if !h.HostsFileLines[idx].IsCommented {
-					h.HostsFileLines[idx].IsCommented = true
-
-					h.HostsFileLines[idx].Raw = h.RenderHostsFileLine(idx)
-					return nil
-				}
-
-				return ErrAlredyCommentedLine
-			}
-		}
-	}
-
-	return ErrHostnameNotFound
-}
-
+// CommentHostsFileLinesByHostname set IsCommented to true on every line that matches a given Hostname
 func (h *HostsFile) CommentHostsFileLinesByHostname(hostname string) {
 	h.Lock()
 	defer h.Unlock()
@@ -694,11 +548,12 @@ func (h *HostsFile) CommentHostsFileLinesByHostname(hostname string) {
 	}
 }
 
-func (h *HostsFile) CommentHostsFileLinesByHostnameAsRegexp(hostname string) {
+// CommentHostsFileLinesByRegexp set IsCommented to true on every line that matches a given regexp
+func (h *HostsFile) CommentHostsFileLinesByRegexp(pattern string) {
 	h.Lock()
 	defer h.Unlock()
 
-	reg := regexp.MustCompile(hostname)
+	reg := regexp.MustCompile(pattern)
 
 	for idx := range h.HostsFileLines {
 		for _, hn := range h.HostsFileLines[idx].Hostnames {
@@ -714,7 +569,7 @@ func (h *HostsFile) CommentHostsFileLinesByHostnameAsRegexp(hostname string) {
 	}
 }
 
-//UncommentHostsFileLineByRow set the IsCommented bit for the given row to false
+// UncommentHostsFileLineByRow set the IsCommented bit for the given row to false
 func (h *HostsFile) UncommentHostsFileLineByRow(row int) error {
 	h.Lock()
 	defer h.Unlock()
@@ -737,27 +592,7 @@ func (h *HostsFile) UncommentHostsFileLineByRow(row int) error {
 	return ErrUnknown
 }
 
-//UncommentHostsFileLineByIP set the IsCommented bit for the given address to false
-func (h *HostsFile) UncommentHostsFileLineByIP(ip net.IP) error {
-	h.Lock()
-	defer h.Unlock()
-
-	for idx, hfl := range h.HostsFileLines {
-		if net.IP.Equal(ip, hfl.Address) {
-			if h.HostsFileLines[idx].IsCommented {
-				h.HostsFileLines[idx].IsCommented = false
-
-				h.HostsFileLines[idx].Raw = h.RenderHostsFileLine(idx)
-				return nil
-			}
-
-			return ErrAlredyUncommentedLine
-		}
-	}
-
-	return ErrNotAnAddressLine
-}
-
+// UncommentHostsFileLinesByIP set IsCommented to false for every line that matches a given net.IP
 func (h *HostsFile) UncommentHostsFileLinesByIP(ip net.IP) {
 	h.Lock()
 	defer h.Unlock()
@@ -773,41 +608,13 @@ func (h *HostsFile) UncommentHostsFileLinesByIP(ip net.IP) {
 	}
 }
 
-//UncommentHostsFileLineByAddress set the IsCommented bit for the given address as string to false
-func (h *HostsFile) UncommentHostsFileLineByAddress(address string) error {
-	ip := net.ParseIP(address)
-
-	return h.UncommentHostsFileLineByIP(ip)
-}
-
+// UncommentHostsFileLinesByAddress set IsCommented to false for every line that matches a given net.IP as String
 func (h *HostsFile) UncommentHostsFileLinesByAddress(address string) {
 	ip := net.ParseIP(address)
 	h.UncommentHostsFileLinesByIP(ip)
 }
 
-//UncommentHostsFileLineByHostname set the IsCommented bit for the given hostname to false
-func (h *HostsFile) UncommentHostsFileLineByHostname(hostname string) error {
-	h.Lock()
-	defer h.Unlock()
-
-	for idx := range h.HostsFileLines {
-		for _, hn := range h.HostsFileLines[idx].Hostnames {
-			if hn == hostname {
-				if h.HostsFileLines[idx].IsCommented {
-					h.HostsFileLines[idx].IsCommented = false
-
-					h.HostsFileLines[idx].Raw = h.RenderHostsFileLine(idx)
-					return nil
-				}
-
-				return ErrAlredyUncommentedLine
-			}
-		}
-	}
-
-	return ErrHostnameNotFound
-}
-
+// UncommentHostsFileLinesByHostname set IsCommented to false for every line that matches a given Hostname
 func (h *HostsFile) UncommentHostsFileLinesByHostname(hostname string) {
 	h.Lock()
 	defer h.Unlock()
@@ -825,11 +632,12 @@ func (h *HostsFile) UncommentHostsFileLinesByHostname(hostname string) {
 	}
 }
 
-func (h *HostsFile) UncommentHostsFileLinesByHostnameAsRegexp(hostname string) {
+// UncommentHostsFileLinesByRegexp set IsCommented to false for every line that matches a given regexp
+func (h *HostsFile) UncommentHostsFileLinesByRegexp(pattern string) {
 	h.Lock()
 	defer h.Unlock()
 
-	reg := regexp.MustCompile(hostname)
+	reg := regexp.MustCompile(pattern)
 
 	for idx := range h.HostsFileLines {
 		for _, hn := range h.HostsFileLines[idx].Hostnames {
