@@ -1,38 +1,45 @@
 package libhosty
 
 import (
-	"runtime"
 	"strings"
 	"testing"
 )
 
-//TestParseHostsFile implicitly tests parser()
+var (
+	// for this test use a custom hosts file
+	customHostsFile = `## Custom hosts file
+		# To check if everything works
+		1.1.1.1		my.cloudflare.domain
+		8.8.8.8		my.google.dns
+		6.6.6.6		an.evil.domain
+		# 12.12.12.12	commented.evil.domain # with comments
+		# 2.3.4.5		first.domain second.domain third.domain
+		5.6.7.8		first.domain second.domain third.domain
+		5.6.7.8		a.domain b.domain c.domain`
+)
+
+// TestParseHostsFile implicitly tests parser()
 // assuming the test is executed on a computer,
 // we should be able to load the default hosts file
 func TestParseHostsFile(t *testing.T) {
-	var path string
-
 	// check for invalid path
-	path = "/my/custom/invalid/path/hosts"
-	_, err := ParseHostsFile(path)
+	path := "/my/custom/invalid/path/hosts"
+	_, err := InitFromCustomPath(path)
 	if err == nil {
 		t.Fatalf("should fail with invalid path: %s", path)
 	}
 
-	// define file path
-	switch runtime.GOOS {
-	case "win":
-		path = "C:\\Windows\\System32\\drivers\\etc\\hosts"
-	default:
-		path = "/etc/hosts"
-	}
-
 	// parse file
-	hf, err := ParseHostsFile(path)
-
+	hf, err := Init()
 	// check for errors
 	if err != nil {
 		t.Fatalf("error parsing hosts file: %s", err)
+	}
+
+	// load custom file for tests
+	hf.HostsFileLines, err = ParseHostsFileFromString(customHostsFile)
+	if err != nil {
+		t.Fatalf("error parsing custom hosts file: %s", err)
 	}
 
 	// check for unknown lines
@@ -49,17 +56,16 @@ func TestParseHostsFile(t *testing.T) {
 	// 	}
 	// }
 
-	// for every address line, ensure we have a valid ip address
-	for k, v := range hf {
+	// for every  line,
+	for k, v := range hf.HostsFileLines {
+		// ensure addresses have a valid ip
 		if v.Type == LineTypeAddress {
 			if v.Address == nil {
 				t.Fatalf("address line without address at index %d: %v", k, v)
 			}
 		}
-	}
 
-	// for every comment line ensure it starts with #
-	for k, v := range hf {
+		// ensure comments starts with #
 		if v.Type == LineTypeComment {
 			if !strings.HasPrefix(v.Raw, "#") {
 				t.Fatalf("comment line does not starts with # at index %d: %v", k, v)
@@ -68,24 +74,29 @@ func TestParseHostsFile(t *testing.T) {
 	}
 }
 
-//TestParseHostsFileAsString test unknown line type
+// TestParseHostsFileFromString test unknown line type
 // actual parsing is already tested in TestParseHostsFile()
-func TestParseHostsFileAsString(t *testing.T) {
-	// define a custom hosts file with an address line and an invalid line"
-	var fakeHostsFile = `1.1.1.1 my.hosts.file # With Comment
-129dj120isdj12i0 1092jd 210dk`
-
-	hf, err := ParseHostsFileAsString(fakeHostsFile)
+func TestParseHostsFileFromString(t *testing.T) {
+	// parse custom hosts file from string
+	hf, err := ParseHostsFileFromString(customHostsFile)
 	if err != nil {
-		t.Fatalf("error parsing fakeHostsFile: %s", err)
+		t.Fatalf("error parsing customHostsFile: %s", err)
 	}
 
-	// we expect the 1st line to be of type address
-	if hf[0].Type != LineTypeAddress {
-		t.Fatalf("line should be of type address: %v", hf[0])
-	}
-	// and the 2nd line to be of type unknown
-	if hf[1].Type != LineTypeUnknown {
-		t.Fatalf("line should be of type address: %v", hf[1])
+	// for every  line,
+	for k, v := range hf {
+		// ensure addresses have a valid ip
+		if v.Type == LineTypeAddress {
+			if v.Address == nil {
+				t.Fatalf("address line without address at index %d: %v", k, v)
+			}
+		}
+
+		// ensure comments starts with #
+		if v.Type == LineTypeComment {
+			if !strings.HasPrefix(v.Raw, "#") {
+				t.Fatalf("comment line does not starts with # at index %d: %v", k, v)
+			}
+		}
 	}
 }
